fix showing non-item items in styles causing error
